refactor: components => frontend/components
