+++ website/prorata/frontend/index.ts
@@ -4,3 +4,4 @@ export * from './use-mounted-effect'
 export * from './use-async-state'
 export * from './use-layout-effect'
 export * from './use-auto-focus'
+export * from './components'
